+++ tests/integration/test_train.py
@@ -120,7 +120,10 @@ def test_metrics(nequip_dataset, BENCHMARK_ROOT, conffile, builder):
         # We just don't add rescaling:
         true_config["model_builders"] = [builder]
         # We need truth labels as inputs for these fake testing models
-        true_config["_override_allow_truth_label_inputs"] = True
+        true_config["model_input_fields"] = {
+            AtomicDataDict.FORCE_KEY: "1o",
+            AtomicDataDict.TOTAL_ENERGY_KEY: "0e",
+        }
 
         config_path = tmpdir + "/conf.yaml"
         with open(config_path, "w+") as fp:
@@ -238,7 +241,10 @@ def test_requeue(nequip_dataset, BENCHMARK_ROOT, conffile):
         # We just don't add rescaling:
         true_config["model_builders"] = [builder]
         # We need truth labels as inputs for these fake testing models
-        true_config["_override_allow_truth_label_inputs"] = True
+        true_config["model_input_fields"] = {
+            AtomicDataDict.FORCE_KEY: "1o",
+            AtomicDataDict.TOTAL_ENERGY_KEY: "0e",
+        }
 
         for irun in range(3):